new file for strings, moved stuff around
